Move Heartbeat class into Utilities directory
